refactor: extract reporting helpers from app.py to porosity_fe.reporting
The pure-Python helpers previously defined inside app.py (parse_layup,
build_export_payload, write_results_json/csv, governing_failure,
recommend_disposition, build_ncr_record, serialise_ncr_json/markdown/pdf,
write_ncr_json/markdown/pdf, plus their private helpers and the
STRUCTURAL_CLASSES constant) now live in porosity_fe/reporting.py. app.py
re-imports them so the Streamlit UI keeps working unchanged.

Why:
- Lets the core test matrix run without streamlit installed. Tests
  previously did `from app import build_ncr_record, ...`, which pulls in
  `import streamlit as st` at app.py module load. With the helpers in a
  dedicated package module, tests now `from porosity_fe.reporting import
  ...` and never touch streamlit. This closes the loop from the
  streamlit-optional split (#155): the optional `web` extra is finally
  only needed by callers who actually launch the Streamlit UI.
- Migrates app.py off the porosity_fe_analysis compatibility shim, in
  line with the documentation migration in #160. Both the top-level
  `from porosity_fe_analysis import ...` and the lazy imports inside the
  extracted helpers now target `porosity_fe` directly.
- matplotlib.pyplot is imported lazily inside serialise_ncr_pdf rather
  than at module top, keeping porosity_fe.reporting cheap to import.

Verified:
- ruff and the full pytest suite (514 passed, 1 skipped) pass both with
  and without streamlit installed.
- `python -c "import app"` still works when streamlit is installed.
- No test file imports `app` anymore.
